requirements.txt
django==4.2.11
flake8==6.1.0
python-dotenv==1.0.1
django-split-settings==1.3.1
django-debug-toolbar==4.3.0
psycopg-binary==3.1.19